// src/pages/dashboard/dashboard.types.ts
export type AggregatedMetricsType = {
    totalRevenue: number;
    avgJobSize: number;
    totalCom: number;
    totalLeads: number;
    costPerLead: number;
    totalAppointmentsSet: number;
    costPerAppointmentSet: number;
    estimateSetRate: number;
    estimateSetCount: number;
    unqualifiedCount: number;
    costPerEstimateSet: number;
}